extracted zetzer helper functions to module
